requirements.txt
streamlit==1.37.1
qrcode==7.4.2
Pillow==10.4.0
